requirements.txt
PyExecJS==1.5.1
requests_html==0.10.0